Add Edit menu, add shortcuts for all actions
The main purpose of this edit menu is to make it clear what actions are
available and what the buttons in the grid do by putting a text label
next to the used icons.

Note that some shortcuts are defined using the ShortcutKeys property on
the menu item of the Edit menu, making them trigger globally.

Others only have the ShortcutKeyDisplayString property and implement the
logic in the OnKeyDown event handler, mostly because it would otherwise
conflict with editing shortcuts like Ctrl+C or Ctrl+Z.