Rename Save to Save As, update about screen, add ellipsis to menu items
The Save operation in the File menu is better called Save As because it
will always show a file dialog rather than replacing the current file.

The about dialog contained a link to the Codeplex website, make it point
directly to Github instead. Fix the copyright years while we're at it.

Finally add an ellipsis to all menu items which open a new dialog. It's
an old convention so that a (knowledgeable) user knows  that the
particular menu item is not an instant action.