Allow Ignore/DictionaryAdd/UseSuggestion actions to work on selected cells
The grid allows for selections to be made by click + dragging across it,
using Shift+Click or picking one by one using Ctrl+Click.

Since all operations now have a shortcut as well as a menu option we can
execute the chosen action on all the rows in which we selected cells for
faster editing.