Only trim common suffixes
Check if a suffix ends with s, t, d or ve for common contractions or
possesive s.

Previously the method would remove anything after an apostrophe, but in
case of missing spaces this might even have been entire words. As a
result, such errors might have been marked as being found in the
dictionary because the prefix was.